Refactor : component 폴더 생성, handleDate 파일 이동, listItems 컴포넌트로 기존 맵핑 분리
